Moved system CLI commands to their own package
