+++ cdap-cli/src/main/java/co/cask/cdap/cli/command/system/SearchCommandsCommand.java
@@ -14,10 +14,11 @@
  * the License.
  */
 
-package co.cask.cdap.cli.command;
+package co.cask.cdap.cli.command.system;
 
 import co.cask.cdap.cli.ArgumentName;
 import co.cask.cdap.cli.CommandCategory;
+import co.cask.cdap.cli.command.system.HelpCommand;
 import co.cask.common.cli.Arguments;
 import co.cask.common.cli.Command;
 import co.cask.common.cli.CommandSet;
@@ -86,7 +87,7 @@ public class SearchCommandsCommand extends HelpCommand {
 
   @Override
   public String getPattern() {
-    return String.format("search commands <%s>", ArgumentName.QUERY);
+    return String.format("cli search commands <%s>", ArgumentName.QUERY);
   }
 
   @Override
