+++ cdap-cli/src/main/java/co/cask/cdap/cli/docs/GenerateCLIDocsTableCommand.java
@@ -17,7 +17,7 @@
 package co.cask.cdap.cli.docs;
 
 import co.cask.cdap.cli.CommandCategory;
-import co.cask.cdap.cli.command.HelpCommand;
+import co.cask.cdap.cli.command.system.HelpCommand;
 import co.cask.common.cli.Arguments;
 import co.cask.common.cli.Command;
 import co.cask.common.cli.CommandSet;
